simplify root pacakges for samples
