--- azure-samples/azure-ad-integration-spring-boot-starter-sample/src/main/java/com/microsoft/azure/spring/autoconfigure/aad/sample/AzureAdIntegrationSpringBootAutoconfigureSampleApplication.java
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for
- * license information.
- */
-package com.microsoft.azure.spring.autoconfigure.aad.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AzureAdIntegrationSpringBootAutoconfigureSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AzureAdIntegrationSpringBootAutoconfigureSampleApplication.class, args);
-    }
-}
