+++ azure-samples/azure-ad-integration-spring-boot-starter-sample/src/main/java/sample/aad/AzureAdIntegrationSpringBootAutoconfigureSampleApplication.java
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for
+ * license information.
+ */
+package sample.aad;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AzureAdIntegrationSpringBootAutoconfigureSampleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AzureAdIntegrationSpringBootAutoconfigureSampleApplication.class, args);
+    }
+}
