--- azure-samples/azure-cloud-foundry-service-sample/src/main/java/com/microsoft/azure/cloudfoundry/service/sample/AzureCloudFoundryServiceSampleApplication.java
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for
- * license information.
- */
-package com.microsoft.azure.cloudfoundry.service.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AzureCloudFoundryServiceSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AzureCloudFoundryServiceSampleApplication.class, args);
-    }
-}
