+++ azure-samples/azure-documentdb-spring-boot-starter-sample/src/main/java/sample/documentdb/User.java
@@ -1,0 +1,62 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for
+ * license information.
+ */
+
+package sample.documentdb;
+
+import com.microsoft.azure.spring.data.documentdb.core.mapping.Document;
+
+@Document(collection = "mycollection")
+public class User {
+    private String id;
+    private String firstName;
+    private String lastName;
+    private String address;
+
+    public User(String id, String firstName, String lastName, String address) {
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.address = address;
+    }
+
+    public String getId() {
+        return this.id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("%s %s, %s", firstName, lastName, address);
+    }
+}
+
